Moved utility functions model and dismantled last global utils module
Moved and refactored the former structure utilities module functions to the io_p3d package utils module
Fixed mistakes in the addon init file that happened during the util module relocations
