--- Arma3ObjectBuilder/utilities/__init__.py
@@ -1,2 +1,0 @@
-# In order of dependency
-from . import structure
